refactor: 챌린지 관련 모달 위치 변경 (shared -> features)
