+++ src/features/challenge/components/modal/camera-modal/styles.ts
