+++ src/features/challenge/components/modal/image-zoom-modal/image-zoom-modal.tsx
