+++ src/features/challenge/components/modal/image-zoom-modal/styles.ts
